+++ qgis_hub_plugin/utilities/common.py
@@ -1,6 +1,7 @@
 import os
 from pathlib import Path
 
+from qgis.core import QgsApplication
 from qgis.PyQt.QtGui import QIcon
 
 from qgis_hub_plugin.__about__ import DIR_PLUGIN_ROOT
@@ -22,3 +23,28 @@ def download_file(url: str, file_path: Path, force: bool = False):
         raise OSError(ex)
 
     return file_path.exists()
+
+
+def shorten_string(text: str) -> str:
+    if len(text) > 20:
+        text = text[:17] + "..."
+    return text
+
+
+def download_resource_thumbnail(url: str, uuid: str):
+    qgis_user_dir = QgsApplication.qgisSettingsDirPath()
+    # Assume it as jpg
+    extension = ".jpg"
+    try:
+        extension = url.split(".")[-1]
+    except IndexError():
+        pass
+
+    thumbnail_dir = Path(qgis_user_dir, "qgis_hub", "thumbnails")
+    thumbnail_path = Path(thumbnail_dir, f"{uuid}.{extension}")
+    if not thumbnail_dir.exists():
+        thumbnail_dir.mkdir(parents=True, exist_ok=True)
+
+    download_file(url, thumbnail_path)
+    if thumbnail_path.exists():
+        return thumbnail_path
